// src/common/helpers/base-query.dto.ts
import z from "zod";

export const QuerySchema = z.object({
  page: z.coerce.number().min(1).default(1),
  limit: z.coerce.number().min(1).max(100).default(50),
  sort: z.string().default("createdAt"),
  order: z.enum(["asc", "desc"]).default("desc"),
  search: z.string().optional().default(""),
});

export type Query = z.infer<typeof QuerySchema>;